Improve docker to build separate images for dev and pro
