+++ .docker/app-dev/extra.ini
@@ -1,0 +1,19 @@
+date.timezone = Europe/Madrid
+
+upload_max_filesize = 20M
+post_max_size = 20M
+
+memory_limit = -1
+opcache.enable = 1
+opcache.revalidate_path = 0
+opcache.max_accelerated_files = 20000
+opcache.validate_timestamps = 1
+opcache.memory_consumption = 256
+realpath_cache_size = 5M
+realpath_cache_ttl = 7200
+
+error_log = /proc/self/fd/2
+
+xdebug.mode = debug
+xdebug.start_with_request = trigger
+xdebug.discover_client_host = 1
