+++ .docker/app-prod/www.conf
@@ -1,0 +1,16 @@
+[www]
+
+user = root
+group = root
+
+listen = 127.0.0.1:9000
+
+pm = dynamic
+pm.max_children = 50
+pm.start_servers = 5
+pm.min_spare_servers = 5
+pm.max_spare_servers = 5
+
+clear_env = no
+catch_workers_output = yes
+decorate_workers_output = no
