+++ .env
@@ -1,0 +1,11 @@
+APP_ENV=dev
+APP_SECRET=308d47268412acc4fe69f0afaf00b3e63cc7c69f6d94c62d4530d045cd911839
+
+MYSQL_DATABASE=drupal
+MYSQL_USER=drupal
+MYSQL_PASSWORD=drupal
+MYSQL_HOST=mysql
+
+GTM_ID=GTM-KTPM543
+COOKIES_DEFAULT_DOMAIN=localhost
+TRUSTED_HOST=localhost
