+++ .gitignore
@@ -8,6 +8,9 @@
 .certs
 
 # DRUPAL
+.env.local
+.env.local.php
+.env.*.local
 .php_cs
 .php_cs.cache
 .phpunit.result.cache
@@ -18,7 +21,10 @@ phpunit.xml
 web/core
 web/modules/contrib
 web/themes/contrib
-web/sites/default/files
+web/sites/default/files/*
+!web/sites/default/files/tmp
+web/sites/default/files/tmp/*
+!web/sites/default/files/tmp/.gitignore
 web/themes/custom/runroom/build/*
 !web/themes/custom/runroom/build/.gitignore
