+++ composer.json
@@ -19,13 +19,13 @@
         "drupal/clientside_validation": "3.0.0-rc4",
         "drupal/config_filter": "1.8.0",
         "drupal/config_split": "1.7.0",
-        "drupal/core-recommended": "9.1.3",
+        "drupal/core-recommended": "9.1.4",
         "drupal/date_popup": "1.1.0",
         "drupal/drush_language": "1.0.0-rc4",
         "drupal/editor_advanced_link": "1.8.0",
         "drupal/entity_reference_revisions": "1.8.0",
         "drupal/field_group": "3.1.0",
-        "drupal/gin": "3.0.0-alpha32",
+        "drupal/gin": "3.0.0-alpha33",
         "drupal/gin_toolbar": "1.0.0-beta14",
         "drupal/hreflang": "1.3.0",
         "drupal/imageapi_optimize": "4.0.0-beta1",
@@ -40,7 +40,9 @@
         "drupal/vbo_export": "3.2.0",
         "drupal/views_bulk_operations": "3.10.0",
         "drupal/xmlsitemap": "1.0.0",
-        "drush/drush": "10.3.6"
+        "drush/drush": "10.3.6",
+        "symfony/dotenv": "5.2.3",
+        "symfony/flex": "1.12.1"
     },
     "replace": {
         "symfony/polyfill-php56": "*",
@@ -62,10 +64,10 @@
         "friendsofphp/php-cs-fixer": "2.18.2",
         "kint-php/kint": "3.3.0",
         "kint-php/kint-twig": "3.0.0",
-        "mglaman/phpstan-drupal": "0.12.7",
+        "mglaman/phpstan-drupal": "0.12.8",
         "phpspec/prophecy-phpunit": "2.0.1",
-        "phpstan/phpstan": "0.12.70",
-        "phpunit/phpunit": "9.5.1"
+        "phpstan/phpstan": "0.12.76",
+        "phpunit/phpunit": "9.5.2"
     },
     "config": {
         "sort-packages": true
