+++ docker-compose.yaml
@@ -1,0 +1,47 @@
+version: '3'
+
+services:
+    app:
+        build:
+            context: .
+            target: fpm-dev
+        volumes:
+            - .:/usr/app/:cached
+            - .docker/app-dev/extra.ini:/usr/local/etc/php/conf.d/extra.ini
+            - .docker/app-dev/www.conf:/usr/local/etc/php-fpm.d/www.conf
+            - /usr/app/node_modules/
+            - /usr/app/web/sites/default/files/tmp/
+            - /usr/app/web/sites/default/files/php/
+            - /usr/app/web/sites/default/files/languages/
+            - /usr/app/web/sites/default/files/translations/
+            - ${SSH_AUTH_SOCK}:${SSH_AUTH_SOCK}
+        depends_on:
+            - mysql
+        environment:
+            - SSH_AUTH_SOCK
+            - XDEBUG_CONFIG
+        command: php-fpm --allow-to-run-as-root
+    nginx:
+        image: nginx:1-alpine
+        ports:
+            - 8080:80
+            - 8443:443
+        volumes:
+            - .docker/nginx/default.conf:/etc/nginx/conf.d/default.conf
+            - .certs:/usr/app/.certs:ro,cached
+            - ./web:/usr/app/web:ro,cached
+        depends_on:
+            - app
+    mysql:
+        image: mysql:5.7
+        environment:
+            - MYSQL_DATABASE
+            - MYSQL_USER
+            - MYSQL_PASSWORD
+            - MYSQL_ROOT_PASSWORD=root
+        ports:
+            - 8336:3306
+    mailhog:
+        image: mailhog/mailhog
+        ports:
+            - 8025:8025
