--- docker/app/Dockerfile
@@ -1,14 +1,0 @@
-FROM php:7.4-fpm-alpine
-
-COPY --from=mlocati/php-extension-installer:latest /usr/bin/install-php-extensions /usr/bin/
-
-RUN install-php-extensions apcu bz2 gd intl opcache pdo_mysql zip pcov xdebug
-RUN apk --no-cache add bash git mysql-client openssh-client rsync patch unzip
-
-ENV PATH="/usr/app/vendor/bin:/usr/app/bin:${PATH}"
-
-RUN mv $PHP_INI_DIR/php.ini-production $PHP_INI_DIR/php.ini
-
-COPY --from=composer:2 /usr/bin/composer /usr/bin/composer
-
-WORKDIR /usr/app
