--- docker/app/www.conf
@@ -1,12 +1,0 @@
-[www]
-
-user = root
-group = root
-
-listen = 127.0.0.1:9000
-
-pm = dynamic
-pm.max_children = 50
-pm.start_servers = 5
-pm.min_spare_servers = 5
-pm.max_spare_servers = 5
