--- docker/docker-compose-dev.yaml
@@ -1,19 +1,0 @@
-version: '3'
-
-services:
-    app:
-        volumes:
-            - ../:/usr/app/:cached
-            - /usr/app/node_modules/
-            - /usr/app/web/sites/default/files/tmp/
-            - /usr/app/web/sites/default/files/php/
-            - /usr/app/web/sites/default/files/languages/
-            - /usr/app/web/sites/default/files/translations/
-            - ${SSH_AUTH_SOCK}:${SSH_AUTH_SOCK}
-        environment:
-            - SSH_AUTH_SOCK
-            - XDEBUG_CONFIG
-    nginx:
-        volumes:
-            - ../.certs:/usr/app/.certs:ro,cached
-            - ../web:/usr/app/web:ro,cached
