--- docker/docker-compose.yaml
@@ -1,32 +1,0 @@
-version: '3'
-
-services:
-    app:
-        build: ./app
-        volumes:
-            - ./app/extra.ini:/usr/local/etc/php/conf.d/extra.ini
-            - ./app/www.conf:/usr/local/etc/php-fpm.d/www.conf
-        command: php-fpm --allow-to-run-as-root
-    nginx:
-        image: nginx:1-alpine
-        ports:
-            - 8080:80
-            - 8443:443
-        volumes:
-            - ./nginx/default.conf:/etc/nginx/conf.d/default.conf
-            - ./nginx/nginx.conf:/etc/nginx/nginx.conf
-        depends_on:
-            - app
-    mysql:
-        image: mysql:5.7
-        environment:
-            MYSQL_DATABASE: drupal
-            MYSQL_USER: drupal
-            MYSQL_PASSWORD: drupal
-            MYSQL_ROOT_PASSWORD: root
-        ports:
-            - 8336:3306
-    mailhog:
-        image: mailhog/mailhog
-        ports:
-            - 8025:8025
