--- docker/nginx/default.conf
@@ -1,102 +1,0 @@
-map $sent_http_content_type $expires {
-    default off;
-
-    image/jpg 1y;
-    image/png 1y;
-    image/gif 1y;
-    image/jpeg 1y;
-    image/svg+xml 1y;
-    image/x-icon 1y;
-    image/vnd.microsoft.icon 1y;
-    application/x-font-woff 1y;
-    font/woff 1y;
-    font/woff2 1y;
-    application/pdf 1y;
-    video/mp4 1y;
-    video/webm 1y;
-    text/plain 1y;
-    text/xml 1y;
-    application/rss+xml 1y;
-    application/xhtml+xml 1y;
-    application/xml 1y;
-    text/x-javascript 1y;
-    text/javascript 1y;
-    application/x-javascript 1y;
-    application/javascript 1y;
-    text/css 1y;
-}
-
-server {
-    listen 80;
-
-    return 301 https://$host:8443$request_uri;
-}
-
-server {
-    listen 443 ssl http2 default_server;
-
-    root /usr/app/web;
-
-    client_max_body_size 0;
-
-    # Local images
-    # location @production {
-    #     set $production https://symfony.runroom.dev;
-    #     proxy_pass $production;
-    #     proxy_hide_header Set-Cookie;
-    # }
-    # location ~* ^/.+\.(jpe?g|gif|png|ico|svg|pdf)$ {
-    #     try_files $uri $uri/ @production;
-    # }
-
-    location / {
-        try_files $uri /index.php$is_args$args;
-    }
-
-    location ~ \.php$ {
-        fastcgi_pass app:9000;
-        fastcgi_split_path_info ^(.+\.php)(/.*)$;
-        include fastcgi_params;
-
-        fastcgi_param SCRIPT_FILENAME $realpath_root$fastcgi_script_name;
-        fastcgi_param DOCUMENT_ROOT $realpath_root;
-    }
-
-    ssl_certificate /usr/app/.certs/certificate.pem;
-    ssl_certificate_key /usr/app/.certs/certificate-key.pem;
-    ssl_session_timeout 1d;
-    ssl_session_cache shared:SSL:50m;
-    ssl_session_tickets off;
-    ssl_protocols TLSv1.3 TLSv1.2;
-    ssl_ciphers ECDHE-ECDSA-AES256-GCM-SHA384:ECDHE-RSA-AES256-GCM-SHA384:ECDHE-ECDSA-CHACHA20-POLY1305:ECDHE-RSA-CHACHA20-POLY1305:ECDHE-ECDSA-AES128-GCM-SHA256:ECDHE-RSA-AES128-GCM-SHA256:ECDHE-ECDSA-AES256-SHA384:ECDHE-RSA-AES256-SHA384:ECDHE-ECDSA-AES128-SHA256:ECDHE-RSA-AES128-SHA256;
-    ssl_prefer_server_ciphers on;
-    ssl_ecdh_curve secp384r1;
-
-    # Needed on production, disabled on local
-    # ssl_stapling on;
-    # ssl_stapling_verify on;
-    # ssl_dhparam /etc/nginx/dhparam.pem; // sudo openssl dhparam -out /etc/nginx/dhparam.pem 4096
-
-    resolver_timeout 5s;
-    resolver 1.1.1.1 1.0.0.1 valid=300s;
-
-    expires $expires;
-    etag off;
-    server_tokens off;
-    http2_push_preload on;
-
-    add_header Strict-Transport-Security "max-age=63072000; includeSubDomains; preload" always;
-    add_header X-Content-Type-Options nosniff always;
-    add_header X-Frame-Options SAMEORIGIN always;
-    add_header X-XSS-Protection "1; mode=block" always;
-
-    gzip on;
-    gzip_disable "msie6";
-    gzip_vary on;
-    gzip_proxied any;
-    gzip_comp_level 6;
-    gzip_buffers 16 8k;
-    gzip_http_version 1.1;
-    gzip_min_length 256;
-    gzip_types application/javascript application/rss+xml application/x-javascript application/xhtml+xml application/xml image/svg+xml image/x-icon image/vnd.microsoft.icon text/css text/javascript text/x-javascript text/plain text/xml;
-}
