--- tests/DemoTest.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests;
-
-use PHPUnit\Framework\TestCase;
-
-class DemoTest extends TestCase
-{
-    /** @test */
-    public function itWorks()
-    {
-        $this->assertTrue(true);
-    }
-}
