+++ tests/Unit/DemoTest.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Unit;
+
+use PHPUnit\Framework\TestCase;
+
+class DemoTest extends TestCase
+{
+    /** @test */
+    public function itWorks()
+    {
+        $this->assertTrue(true);
+    }
+}
