--- web/sites/custom.services.yml
@@ -1,9 +1,0 @@
-parameters:
-  http.response.debug_cacheability_headers: true
-  twig.config:
-    debug: true
-    auto_reload: true
-    cache: true
-services:
-  cache.backend.null:
-    class: Drupal\Core\Cache\NullBackendFactory
